pyproject.toml
[tool.poetry]
name = "hop3"
version = "0.1.0"
description = ""
authors = ["Stefane Fermigier <sf@abilian.com>"]
readme = "README.md"

packages = [
    { include = "hop3", from = "src" }
]

[tool.poetry.scripts]
hop-agent = "hop3.main:main"
hop-test = "hop3.testing.main:main"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
cleez = "*"
pluggy = "^1.4.0"

# Remove later
devtools = "^0.12.2"
sqlalchemy = "^2.0.29"
sqlalchemy-serializer = "^1.4.12"
advanced-alchemy = "^0.9"
psycopg2 = "^2.9.9"
snoop = "^0.4.3"
svcs = "^24.1.0"

[tool.poetry.group.dev.dependencies]
abilian-devtools = "^0.5.18"
pytest = "^8.0.0"
pyinfra = "^2.9"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
infer-types = "^1.0.0"
types-paramiko = "^3.4.0.20240311"
reuse = "^3.0.1"
types-psycopg2 = "^2.9.21.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pyright]
exclude = [
    "src/hop3/util/templating.py",
]
